struct optimizations wip

+++ src/main/java/de/linusdev/lutils/nat/abi/ABI.java
@@ -82,7 +82,7 @@ public interface ABI {
             boolean compress,
             @NotNull MemorySizeable children,
             int length,
-            int stride
+            long stride
     );
 
     /**
